requirements.txt
streamlit
ultralytics==8.0.200
opencv-python-headless==4.8.1.78
numpy
Pillow